Fix: Set ALC_MAX_AUXILIARY_SENDS to 4
Old mods expect 4 sends slots because it's been that way for so long.
https://github.com/djpadbit/Sound-Physics/blob/4641cd185310a1e3e809f062e424e47165954f9f/src/main/java/com/sonicether/soundphysics/SoundPhysics.java#L184-L199

The OpenAL we use for now matches the one used by lwjgl 3.3.3, which
sets the default to 2.

https://github.com/kcat/openal-soft/commit/d3cc867bd42759cb8e294d691354187984f96ff4

We set it back to 4 so old mods function properly, there is pretty much
no downside to doing so since it was being set to the max just a while
ago
https://github.com/kcat/openal-soft/commit/69be02c0fdba3cda8319a6299de1d45b5c695735

When testing, for some reason we max out at 6 sends instead of 16. Not
like mods can expect anything else unless they override Paul's Sound
System...